+++ app/src/main/java/eu/faircode/email/FragmentOptions.java
@@ -132,7 +132,7 @@ public class FragmentOptions extends FragmentBase {
             "startup", "cards", "beige", "tabular_card_bg", "shadow_unread", "shadow_highlight",
             "portrait2", "portrait2c", "portrait_min_size", "landscape", "landscape_min_size",
             "nav_count", "nav_unseen_drafts", "navbar_colorize",
-            "indentation", "date", "date_fixed", "date_bold", "threading", "threading_unread",
+            "indentation", "group_category", "date", "date_fixed", "date_bold", "threading", "threading_unread",
             "highlight_unread", "highlight_color", "color_stripe", "color_stripe_wide",
             "avatars", "bimi", "gravatars", "favicons", "generated_icons", "identicons", "circular", "saturation", "brightness", "threshold",
             "authentication", "authentication_indicator",